requirements.txt
# Requirements for the cycle-refactoring agentic pipeline
# Edit versions as appropriate for your environment.

langchain>=0.0.200
langchain-ollama>=0.1.0
langchain-community>=0.1.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.1.0
chromadb>=0.3.24
pydantic>=1.10
PyYAML>=6.0
scikit-learn>=1.2
regex>=2023.8.8
requests>=2.28
pypdf>=3.0

# Optional / dev
# anthropic - REMOVED for air-gapped environment
